update to master_combined.py
Added a zip() function to combine a business name list with a rating list, and print the results to the screen.  I was able to pull out the hyphens in the business names with .replace("-"," ").

Figured out that the root of the "can't convert to float" is likely a result of google not being able to parse the user's input into a real address (typically a not-huge business name).

Considering requiring state, as there are multiple locations for landmark names (e.g. Pioneer Square exists in Portland, OR as well as Seattle, WA, and google defaults to Seattle when I want to search from the Portland location)